Align NVFP4 merge arithmetic and FSDP2 state handling
Qualify routed FSDP2 RAM loading and CPU offload, preserve FP32 merge arithmetic across native, expert and TP paths, and add focused Blackwell NVFP4 CI with H100 defaults retained.

+++ src/axolotl/monkeypatch/torchao_nvfp4_merge.py
@@ -112,10 +112,8 @@ def quantize_native_effective_weight(base_weight, lora_a, lora_b, scaling):
     """Encode ``dequant(base) + scaling * (B @ A)`` with the base's native recipe."""
     recipe = capture_native_nvfp4_recipe(base_weight)
     with torch.autocast(device_type=lora_a.device.type, enabled=False):
-        delta = (lora_b @ lora_a) * scaling
-        effective = (base_weight.dequantize().float() + delta.float()).to(
-            recipe.orig_dtype
-        )
+        delta = (lora_b.float() @ lora_a.float()) * scaling
+        effective = (base_weight.dequantize().float() + delta).to(recipe.orig_dtype)
     return recipe.quantize(effective)
